TIK-44 adding new right click items to the subproject tree and task tree views.

+++ tik_manager4/ui/mcv/subproject_tree.py
@@ -286,6 +286,8 @@ class TikSubView(QtWidgets.QTreeView):
         right_click_menu = QtWidgets.QMenu()
         act_new_sub = right_click_menu.addAction(self.tr("New Sub-Project"))
         act_new_sub.triggered.connect(lambda _, x=item: self.new_sub_project(item))
+        act_edit_sub = right_click_menu.addAction(self.tr("Edit Sub-Project"))
+        act_edit_sub.triggered.connect(lambda _, x=item: self.edit_sub_project(item))
         act_delete_sub = right_click_menu.addAction(self.tr("Delete Sub-Project"))
         act_delete_sub.triggered.connect(lambda _, x=item: self.delete_sub_project(item))
         # act_new_sub.triggered.connect(partial(self.TreeItem_Add, level, mdlIdx))
@@ -305,12 +307,25 @@ class TikSubView(QtWidgets.QTreeView):
             message, title = self.model.project.log.get_last_message()
             self._feedback.pop_info(title.capitalize(), message)
 
-    def new_task(self, item):
+    def edit_sub_project(self, item):
         # first check for the user permission:
         if self.model.project._check_permissions(level=2) != -1:
-            _dialog = NewTask(self.model.project, parent_sub=item.subproject, parent=self)
-            state = _dialog.exec_()
-            if state:
+            pass
+            # TODO implement the edit subproject dialog
+            # _dialog = EditSubproject(item.subproject, parent=self)
+            # state = _dialog.exec_()
+            # if state:
+            #     self.model.update_sub(item)
+        else:
+            message, title = self.model.project.log.get_last_message()
+            self._feedback.pop_info(title.capitalize(), message)
+
+    def new_task(self, item):
+        # first check for the user permission:
+        # if self.model.project._check_permissions(level=2) != -1:
+        _dialog = NewTask(self.model.project, parent_sub=item.subproject, parent=self)
+        state = _dialog.exec_()
+        if state:
                 # emit clicked signal
                 # self.item_selected.emit([_dialog.get_created_task()])
                 self.add_item.emit(_dialog.get_created_task())
@@ -320,12 +335,33 @@ class TikSubView(QtWidgets.QTreeView):
             self._feedback.pop_info(title.capitalize(), message)
 
     def delete_sub_project(self, item):
-        state = self.model.project.delete_sub_project(uid=item.subproject.id)
-        if state != -1:
-            self.model.removeRow(item.row(), item.parent().index())
+        # Prompt the user to confirm the deletion
+        message = "Are you sure you want to delete the sub-project: {}?".format(item.subproject.name)
+        title = "Delete Sub-Project"
+        sure = self._feedback.pop_question(title, message, buttons=["ok", "cancel"])
+        if sure == "ok":
+            # double check if the sub-project is not empty
+            if not item.subproject.is_subproject_empty(item.subproject):
+                message = "The sub-project is not empty.\n\nALL TASKS AND SUB-PROJECTS WILL BE DELETED!\n\nARE YOU " \
+                          "SURE YOU WANT TO CONTINUE? "
+                title = "NON EMPTY SUB-PROJECT"
+                really_sure = self._feedback.pop_question(title, message, buttons=["ok", "cancel"])
+                if really_sure != "ok":
+                    return
+            state = self.model.project.delete_sub_project(uid=item.subproject.id)
+            if state != -1:
+                self.model.removeRow(item.row(), item.parent().index())
+
+                # after removing the row, find the current selected one and emit the clicked signal
+                self.get_tasks(self.currentIndex())
+
+            else:
+                message, title = self.model.project.log.get_last_message()
+                self._feedback.pop_info(title.capitalize(), message)
         else:
-            message, title = self.model.project.log.get_last_message()
-            self._feedback.pop_info(title.capitalize(), message)
+            return
+
+
 
 
 class ProxyModel(QtCore.QSortFilterProxyModel):
